+++ lib/color_extensions.dart
@@ -10,18 +10,18 @@ class ColorMisc {
   ///This class Applies values to colors and return either a darker shade or lighter tone of the same color
 
   ///shade_factor < 1.0
-  static Color withShade(Color color, num shade_factor) {
-    var r = color.red * (1 - shade_factor);
-    var g = color.green * (1 - shade_factor);
-    var b = color.blue * (1 - shade_factor);
+  static Color withShade(Color color, num shadeFactor) {
+    var r = color.red * (1 - shadeFactor);
+    var g = color.green * (1 - shadeFactor);
+    var b = color.blue * (1 - shadeFactor);
     return Color.fromRGBO(r.toInt(), g.toInt(), b.toInt(), color.opacity);
   }
 
   ///tint_factor < 1.0
-  static Color withTint(Color color, num tint_factor) {
-    var r = color.red + (255 - color.red) * tint_factor;
-    var g = color.green + (255 - color.green) * tint_factor;
-    var b = color.blue + (255 - color.blue) * tint_factor;
+  static Color withTint(Color color, num tintFactor) {
+    var r = color.red + (255 - color.red) * tintFactor;
+    var g = color.green + (255 - color.green) * tintFactor;
+    var b = color.blue + (255 - color.blue) * tintFactor;
     return Color.fromRGBO(r.toInt(), g.toInt(), b.toInt(), color.opacity);
   }
 
@@ -29,6 +29,6 @@ class ColorMisc {
     assert(name.length > 1);
     final int hash = name.hashCode & 0xffff;
     final double hue = (360.0 * hash / (1 << 15)) % 360.0;
-    return new HSVColor.fromAHSV(1.0, hue, 0.4, 0.90).toColor();
+    return  HSVColor.fromAHSV(1.0, hue, 0.4, 0.90).toColor();
   }
 }